+++ modules/core/src/main/java/org/apache/ignite/IgniteClientDisconnectedException.java
