+++ modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryReaderExImpl.java
@@ -38,6 +38,7 @@ import org.apache.ignite.internal.binary.streams.BinaryInputStream;
 import org.apache.ignite.internal.util.CommonUtils;
 import org.apache.ignite.internal.util.typedef.internal.S;
 import org.apache.ignite.internal.util.typedef.internal.SB;
+import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.marshaller.Marshallers;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -1999,7 +2000,7 @@ class BinaryReaderExImpl implements BinaryReaderEx {
                     throw new BinaryObjectException("Cannot find metadata for object with compact footer " +
                         "(Ignite work directory might have been cleared after restart. Make sure that IGNITE_HOME " +
                         "does not point to a temp folder or any other folder that is destroyed/cleared on restarts) [" +
-                        "typeId=" + typeId + ", IGNITE_HOME='" + CommonUtils.getIgniteHome() + "']");
+                        "typeId=" + typeId + ", IGNITE_HOME='" + U.getIgniteHome() + "']");
 
                 Collection<BinarySchema> existingSchemas = meta.schemas();
